service worker es modules + service worker now uses es modules instead of global objects + fixed bug that caused app to report an available update when offline
