+++ .gitignore
@@ -1,2 +1,2 @@
 .vscode
-debug-latest-commit.txt
+test/
